Combine ConvertToAudio and CreateNewFile into a single Lambda
