--- src/ConvertToAudio/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "convert-to-audio",
-  "version": "1.0.0",
-  "dependencies": {
-    "aws-sdk": "~2",
-    "utils": "file:../../../src/utils"
-  }
-}
