+++ src/CreateNewFile/package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "dependencies": {
     "aws-sdk": "~2",
-    "uuid": "^3.3.2"
+    "uuid": "^3.3.2",
+    "utils": "file:../../../src/utils"
   }
 }
